# openapi.yaml
openapi: 3.0.3
info:
  title: SIEM API
  description: Security Information and Event Management API
  contact:
    name: SIEM Team
    email: siem@company.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  version: 1.0.0
servers:
- url: http://localhost:8080/api
  description: Local development server
- url: https://api.siem.company.com/api
  description: Production server
paths:
  /api/v1/log_sources:
    get:
      tags:
      - log-sources
      summary: List all log sources for the tenant
      description: |-
        Retrieves all log sources configured for the authenticated user's tenant.
        Requires Admin role permissions.
      operationId: list_log_sources
      responses:
        '200':
          description: List of log sources
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogSourceListResponse'
        '403':
          description: Insufficient permissions
        '500':
          description: Internal server error
      security:
      - bearer_auth: []
    post:
      tags:
      - log-sources
      summary: Create a new log source configuration
      description: |-
        Creates a new log source configuration for the authenticated tenant.
        Requires Admin role permissions.
      operationId: create_log_source
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateLogSourceRequest'
        required: true
      responses:
        '201':
          description: Log source created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateLogSourceResponse'
        '400':
          description: Invalid request data
        '403':
          description: Insufficient permissions
        '409':
          description: Log source with this IP already exists
        '500':
          description: Internal server error
      security:
      - bearer_auth: []
...
components:
  schemas:
...
    CreateLogSourceRequest:
      type: object
      required:
      - source_name
      - source_type
      - source_ip
      properties:
        source_ip:
          type: string
        source_name:
          type: string
        source_type:
          type: string
    CreateLogSourceResponse:
      type: object
      required:
      - source_id
      - message
      properties:
        message:
          type: string
        source_id:
          type: string
...
    LogSource:
      type: object
      required:
      - source_id
      - tenant_id
      - source_name
      - source_type
      - source_ip
      - created_at
      properties:
        created_at:
          type: integer
          format: int32
          minimum: 0
        source_id:
          type: string
        source_ip:
          type: string
        source_name:
          type: string
        source_type:
          type: string
        tenant_id:
          type: string
...
    LogSourceListResponse:
      type: object
      required:
      - log_sources
      - total
      properties:
        log_sources:
          type: array
          items:
            $ref: '#/components/schemas/LogSource'
        total:
          type: integer
          minimum: 0
...
  securitySchemes:
    bearer_auth:
      type: http
      scheme: bearer
...